+++ components/services/bat_ledger/in_process/in_process_ledger_service.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/services/bat_ledger/public/cpp/in_process_ledger_service.h"
+#include "brave/components/services/bat_ledger/in_process/in_process_ledger_service.h"
 
 #include <utility>
